backend/requirements.txt
django==5.0.14
djangorestframework==3.15.2
django-environ==0.11.2
django-tenants==3.6.1
psycopg2-binary==2.9.9
drf-spectacular==0.27.2
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.1
celery==5.3.1
django-redis==5.4.0
